+++ MyWeatherApp/Model/ForecastModel/WindForecast.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct WindForecast: Codable {
     
-    var speed: Float
+    var speed: Double
     var degree: Int
     
     private enum CodingKeys: String, CodingKey {
@@ -21,7 +21,7 @@ struct WindForecast: Codable {
     init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         
-        self.speed = try container.decode(Float.self, forKey: .speed)
+        self.speed = try container.decode(Double.self, forKey: .speed)
         self.degree = try container.decode(Int.self, forKey: .degree)
     }
 }
